+++ internal/cloud/fuga/fuga.go
@@ -2,17 +2,17 @@ package fuga
 
 import (
 	"github.com/minectl/internal/automation"
-	"github.com/minectl/internal/cloud/openshift"
+	"github.com/minectl/internal/cloud/openstack"
 )
 
 type Fuga struct {
-	openshift *openshift.Openshift
+	openshift *openstack.OpenStack
 }
 
 const imageName = "Ubuntu 20.04 LTS"
 
 func NewFuga() (*Fuga, error) {
-	client, err := openshift.NewOpenshift(imageName)
+	client, err := openstack.NewOpenStack(imageName)
 	if err != nil {
 		return nil, err
 	}
